+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/util/PackageHelperTest.kt
@@ -1,6 +1,6 @@
 package com.lemonappdev.konsist.core.util
 
-import com.lemonappdev.konsist.util.PackageHelper
+import com.lemonappdev.konsist.core.util.PackageHelper
 import org.amshove.kluent.assertSoftly
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
