+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoFileDeclarationImpl.kt
@@ -4,7 +4,7 @@ import com.lemonappdev.konsist.api.KoModifier
 import com.lemonappdev.konsist.api.declaration.KoFileDeclaration
 import com.lemonappdev.konsist.core.cache.KoDeclarationCache
 import com.lemonappdev.konsist.core.declaration.provider.KoDeclarationProviderUtil
-import com.lemonappdev.konsist.util.PackageHelper
+import com.lemonappdev.konsist.core.util.PackageHelper
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.KtImportDirective
 import org.jetbrains.kotlin.psi.KtImportList
