+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoPsiDeclarationImpl.kt
@@ -2,7 +2,7 @@ package com.lemonappdev.konsist.core.declaration
 
 import com.intellij.psi.PsiElement
 import com.lemonappdev.konsist.api.declaration.KoPsiDeclaration
-import com.lemonappdev.konsist.util.PackageHelper
+import com.lemonappdev.konsist.core.util.PackageHelper
 import org.jetbrains.kotlin.psi.psiUtil.getTextWithLocation
 import java.io.File
