+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/ext/FileExt.kt
@@ -1,8 +1,8 @@
 package com.lemonappdev.konsist.core.ext
 
-import com.lemonappdev.konsist.util.KotlinFileParser
-import com.lemonappdev.konsist.util.KotlinFileParser.KOTLIN_FILE_EXTENSION
-import com.lemonappdev.konsist.util.KotlinFileParser.KOTLIN_SNIPPET_FILE_EXTENSION
+import com.lemonappdev.konsist.core.util.KotlinFileParser
+import com.lemonappdev.konsist.core.util.KotlinFileParser.KOTLIN_FILE_EXTENSION
+import com.lemonappdev.konsist.core.util.KotlinFileParser.KOTLIN_SNIPPET_FILE_EXTENSION
 import java.io.File
 
 internal val File.isKotlinFile get() = isFile && name.endsWith(KOTLIN_FILE_EXTENSION)
